+++ src/main/java/org/opensearch/neuralsearch/stats/events/EventStatName.java
@@ -35,7 +35,26 @@ public enum EventStatName implements StatName {
         "semantic_highlighting_request_count",
         "semantic_highlighting",
         EventStatType.TIMESTAMPED_EVENT_COUNTER
-    );
+    ),
+    // Misc processors
+    TEXT_IMAGE_EMBEDDING_PROCESSOR_EXECUTIONS(
+        "text_image_embedding_executions",
+        "processors.ingest",
+        EventStatType.TIMESTAMPED_EVENT_COUNTER
+    ),
+    SPARSE_ENCODING_PROCESSOR_EXECUTIONS("sparse_encoding_executions", "processors.ingest", EventStatType.TIMESTAMPED_EVENT_COUNTER),
+    NEURAL_QUERY_ENRICHER_PROCESSOR_EXECUTIONS(
+        "neural_query_enricher_executions",
+        "processors.search",
+        EventStatType.TIMESTAMPED_EVENT_COUNTER
+    ),
+    NEURAL_SPARSE_TWO_PHASE_PROCESSOR_EXECUTIONS(
+        "neural_sparse_two_phase_executions",
+        "processors.search",
+        EventStatType.TIMESTAMPED_EVENT_COUNTER
+    ),
+    RERANK_BY_FIELD_PROCESSOR_EXECUTIONS("rerank_by_field_executions", "processors.search", EventStatType.TIMESTAMPED_EVENT_COUNTER),
+    RERANK_ML_PROCESSOR_EXECUTIONS("rerank_ml_executions", "processors.search", EventStatType.TIMESTAMPED_EVENT_COUNTER),;
 
     private final String nameString;
     private final String path;